@@ -2,7 +2,7 @@
   "name": "bedrock-karma",
   "version": "2.2.1-0",
   "description": "Bedrock Karma-based Test Framework",
-  "main": "./lib/test",
+  "main": "./lib/index.js",
   "scripts": {
     "lint": "eslint lib"
   },
@@ -33,6 +33,7 @@
     "babel-runtime": "^6.23.0",
     "chai": "^4.2.0",
     "css-loader": "^0.28.11",
+    "esm": "^3.2.25",
     "istanbul-instrumenter-loader": "^3.0.1",
     "karma": "^3.0.0",
     "karma-babel-preprocessor": "^7.0.0",
@@ -51,12 +52,14 @@
     "webpack": "^4.11.1"
   },
   "devDependencies": {
-    "eslint": "^6.8.0",
-    "eslint-config-digitalbazaar": "^2.3.0"
+    "eslint": "^7.32.0",
+    "eslint-config-digitalbazaar": "^2.8.0",
+    "eslint-plugin-jsdoc": "^37.9.7",
+    "jsdoc-to-markdown": "^6.0.1"
   },
   "peerDependencies": {
-    "bedrock": "^3.0.0",
-    "bedrock-test": "^5.1.0"
+    "bedrock": "^4.5.0",
+    "bedrock-test": "^6.1.0"
   },
   "directories": {
     "lib": "./lib"
