@@ -1,14 +1,13 @@
-/*
- * Copyright (c) 2012-2020 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2012-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const bedrock = require('bedrock');
-const brTest = require('bedrock-test');
-const Server = require('karma').Server;
+import * as bedrock from 'bedrock';
+import * as brTest from 'bedrock-test';
+import httpModule from './http.js';
+import {Server} from 'karma';
 
 // load config defaults
-require('./config');
+import './config.js';
 
 bedrock.events.on('bedrock.tests.run', state => {
   if(brTest.test.shouldRunFramework('karma')) {
@@ -32,7 +31,7 @@ function run(state) {
 
     if(!('httpModule' in karmaConfig)) {
       // override httpModule to handle cluster issues
-      karmaConfig.httpModule = require('./http');
+      karmaConfig.httpModule = httpModule;
     }
 
     const server = new Server(karmaConfig);
