@@ -1,11 +1,12 @@
-/*
- * Copyright (c) 2012-2020 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2012-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import {config} from 'bedrock';
+import {fileURLToPath} from 'url';
+import fs from 'fs';
+import path from 'path';
 
-const config = require('bedrock').config;
-const fs = require('fs');
-const path = require('path');
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const DEFAULT_FRAMEWORKS = ['mocha', 'chai'];
 const DEFAULT_PREPROCESSORS = ['babel', 'webpack'];
