@@ -1,12 +1,10 @@
-/*
- * Copyright (c) 2012-2020 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2012-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import https from 'https';
 
-const https = require('https');
-
-const api = Object.assign({}, https);
-module.exports = api;
+const api = {...https};
+export {api as default};
 
 api.createServer = (...args) => {
   const server = https.createServer.apply(this, args);
